pyproject.toml
[project]
name = "mkdocs-entangled-plugin"
version = "0.6.0"
description = "Plugin for MkDocs helping with rendering Entangled (entangled.github.io) projects."
readme = "README.md"
requires-python = ">=3.12,<4"
license = "Apache-2.0"
authors = [{ name = "Johan Hidding", email = "j.hidding@esciencecenter.nl" }]
dependencies = [
  "entangled-cli>=2.4,<3",
  "mawk>=0.1.7,<0.2",
  "mkdocs>=1.4.2,<2",
]

[project.entry-points."mkdocs.plugins"]
entangled = "mkdocs_entangled:EntangledPlugin"

[dependency-groups]
dev = [
  "mkdocs-material>=9.1.6,<10",
  "mypy>=1.2.0,<2",
  "pyright>=1.1.406",
  "pytest>=7.3.0,<8",
  "pytest-mypy>=0.10.3,<0.11",
]

[tool.hatch.build.targets.sdist]
include = ["mkdocs_entangled"]

[tool.hatch.build.targets.wheel]
include = ["mkdocs_entangled"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = ["--mypy"]
testpaths = ["test"]

[tool.mypy]
packages = ["mkdocs_entangled"]

# [tool.uv.sources]
# entangled-cli = { git = "https://github.com/entangled/entangled.py" }
